Add packages Duke and Exception
