+++ src/main/java/Duke/Disk.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
