+++ src/main/java/Duke/Exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package Duke.Exceptions;
 
 public class DukeException extends Exception{
     public DukeException(String exceptionMessage) {
