+++ src/main/java/Duke/Exceptions/EmptyMessageException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package Duke.Exceptions;
 
 public class EmptyMessageException extends DukeException{
