+++ src/main/java/Duke/Exceptions/WrongCommandException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package Duke.Exceptions;
 
 public class WrongCommandException extends DukeException{
     public WrongCommandException(String exceptionMessage) {
