+++ src/main/java/Duke/Printer.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
